main 브랜치 병합

+++ monicar-control-center/src/main/java/org/controlcenter/vehicle/presentation/dto/VehicleEngineStatusResponse.java
@@ -4,8 +4,8 @@ import lombok.Builder;
 
 @Builder
 public record VehicleEngineStatusResponse(
-	Integer allVehicles,
-	Integer engineOnVehicles,
-	Integer engineOffVehicles
+	Long allVehicles,
+	Long engineOnVehicles,
+	Long engineOffVehicles
 ) {
 }
